read description
- created middleware for exceptions managment

- created repositories user, servicio, reserva and cliente

- created migrations

- update readme

- create data seeds

- updated docker file and created docker compose

+++ Repositories/ServicioRepository.cs
@@ -1,4 +1,5 @@
 ﻿using BookingServiceBackend.Data;
+using BookingServiceBackend.Exceptions;
 using BookingServiceBackend.Models;
 using Microsoft.EntityFrameworkCore;
 
@@ -18,14 +19,14 @@ namespace BookingServiceBackend.Repositories
             try
             {
                 if (await _context.Companies.FirstOrDefaultAsync(c => c.Id == servicio.CompanyId) == null)
-                    throw new Exception("Compañía no encontrada");
+                    throw new NotFoundException("Compañía no encontrada");
 
                 _context.Servicios.Add(servicio);
                 await _context.SaveChangesAsync();
             }
             catch (Exception ex)
             {
-                throw new Exception("Error al agregar el servicio", ex);
+                throw new BadRequestException("Error al agregar el servicio: " + ex.Message);
             }
         }
 
@@ -33,13 +34,18 @@ namespace BookingServiceBackend.Repositories
         {
             try
             {
-                var servicio = await _context.Servicios.FirstOrDefaultAsync(s => s.ServicioId == servicioId && s.CompanyId == companyId) ?? throw new Exception("Servicio no encontrado");
+                var servicio = await _context.Servicios.FirstOrDefaultAsync(s => s.ServicioId == servicioId && s.CompanyId == companyId) ?? throw new NotFoundException("Servicio no encontrado");
+                
+                if (await _context.Reservas.AnyAsync(r => r.ServicioId == servicioId && r.Servicio.CompanyId == companyId))
+                {
+                    throw new BadRequestException("No se puede eliminar el servicio porque tiene reservas asociadas");
+                }
                 _context.Servicios.Remove(servicio);
                 await _context.SaveChangesAsync();
             }
             catch (Exception ex)
             {
-                throw new Exception("Error al eliminar el servicio", ex);
+                throw new BadRequestException("Error al eliminar el servicio: " + ex.Message);
             }
         }
 
@@ -48,11 +54,11 @@ namespace BookingServiceBackend.Repositories
             try
             {
                 var servicio = await _context.Servicios.FirstOrDefaultAsync(s => s.ServicioId == servicioId && s.CompanyId == companyId);
-                return servicio ?? throw new Exception("Servicio no encontrado");
+                return servicio ?? throw new NotFoundException("Servicio no encontrado");
             }
             catch (Exception ex)
             {
-                throw new Exception("Error al obtener el servicio", ex);
+                throw new BadRequestException("Error al obtener el servicio: " + ex.Message);
             }
         }
 
@@ -64,7 +70,7 @@ namespace BookingServiceBackend.Repositories
             }
             catch (Exception ex)
             {
-                throw new Exception("Error al obtener los servicios", ex);
+                throw new BadRequestException("Error al obtener los servicios: " + ex.Message);
             }
         }
 
@@ -72,14 +78,14 @@ namespace BookingServiceBackend.Repositories
         {
             try
             {
-                var servicioToUpdate = await _context.Servicios.FirstOrDefaultAsync(s => s.ServicioId == servicio.ServicioId && s.CompanyId == companyId) ?? throw new Exception("Servicio no encontrado");
+                var servicioToUpdate = await _context.Servicios.FirstOrDefaultAsync(s => s.ServicioId == servicio.ServicioId && s.CompanyId == companyId) ?? throw new NotFoundException("Servicio no encontrado");
                 servicioToUpdate.Nombre = servicio.Nombre;
                 servicioToUpdate.Precio = servicio.Precio;
                 await _context.SaveChangesAsync();
             }
             catch (Exception ex)
             {
-                throw new Exception("Error al actualizar el servicio", ex);
+                throw new BadRequestException("Error al actualizar el servicio: " + ex.Message);
             }
         }
     }
